requirement.txt
tensorflow-gpu
opencv-python
numpy
